Fix for empty sequence #84
When the pickle is valid, but has no known images, then there
will be an empty set of target users, and when checking the
distance to users, it will break with an empty sequence
error as found in #84. This patch is a simple fix that defaults
the loop to the unknown user - no users mean it will consider
it a detected face, but unknown.